requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
ants==0.0.7
antspyx==0.6.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0nnnnn
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bids-validator==1.14.7.post0
bidsschematools==1.0.5
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
deepdiff==8.4.2
defusedxml==0.7.1
dipy==1.11.0
Django==5.2.7
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.1
etelemetry==0.3.1
executing==2.2.0
fastjsonschema==2.21.1
fileformats==0.15.2
filelock==3.18.0
flexcache==0.3
flexparser==0.4
flow_matching==1.0.10
fonttools==4.56.0
formulaic==1.1.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.44
googledrivedownloader==1.1.0
greenlet==3.2.2
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.1.post0
lightning-utilities==0.14.3
lpips==0.1.4
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
monai==1.4.0
monai-generative==0.2.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.5.0
narwhals==1.39.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nilearn==0.12.1
notebook==7.3.3
notebook_shim==0.2.4
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
orderly-set==5.3.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
pimms==0.3.25
Pint==0.24.4
platformdirs==4.3.6
plotly==5.12.0
plyfile==1.1.3
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyAFQ==2.0.1
pybids==0.19.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydra==0.25
Pygments==2.19.1
pyparsing==3.2.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==2.5.1.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
ray==2.46.0
rdflib==7.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.23.1
setproctitle==1.3.5
setuptools-scm==5.0.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.13.1
templateflow==24.2.2
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
toml==0.10.2
torch==2.5.1
torch_geometric==1.3.1
torchaudio==2.5.1
torchdiffeq==0.2.5
torchinfo==1.8.0
torchmetrics==1.7.3
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
trx-python==0.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.6.0
yarl==1.20.1
zope.event==6.0
zope.interface==8.0.1
